@@ -21,20 +21,20 @@
         "ext-curl": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
-        "spatie/ignition": "^1.0",
-        "symfony/console": "^6.0",
-        "symfony/var-dumper": "^6.0",
-        "illuminate/support": "^8.77|^9.22|^10.0"
+        "spatie/ignition": "^1.4.3",
+        "symfony/console": "^6.2.3",
+        "symfony/var-dumper": "^6.2.3",
+        "illuminate/support": "^10.0"
     },
     "require-dev": {
-        "filp/whoops": "^2.14",
-        "livewire/livewire": "^2.0",
-        "mockery/mockery": "^1.4",
-        "orchestra/testbench": "^6.23|^7.0|^8.0",
-        "pestphp/pest": "^1.20",
-        "phpstan/extension-installer": "^1.1",
-        "phpstan/phpstan-deprecation-rules": "^1.0",
-        "phpstan/phpstan-phpunit": "^1.0"
+        "filp/whoops": "^2.14.6",
+        "livewire/livewire": "^2.11",
+        "mockery/mockery": "^1.5.1",
+        "orchestra/testbench": "^8.0",
+        "pestphp/pest": "^1.22.3",
+        "phpstan/extension-installer": "^1.2",
+        "phpstan/phpstan-deprecation-rules": "^1.1.1",
+        "phpstan/phpstan-phpunit": "^1.3.3"
     },
     "config": {
         "sort-packages": true,
